refactor the dist code and check some infomation.
