+++ python/paddle/distributed/communication/__init__.py
@@ -1,0 +1,55 @@
+#   Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from .all_gather import all_gather
+from .alltoall import alltoall, alltoall_single
+from .batch_isend_irecv import batch_isend_irecv
+from .reduce import reduce, ReduceOp
+from .group import new_group, get_group, is_initialized, destroy_process_group
+from .scatter import scatter
+from .split import split
+from .wait import wait
+from .all_reduce import all_reduce
+from .barrier import barrier
+from .broadcast import broadcast
+from .recv import recv, irecv
+from .send import send, isend
+from .reduce_scatter import reduce_scatter
+from .p2p import P2POp
+
+
+__all__ = [
+           "all_gather",
+           "alltoall",
+           "alltoall_single",
+           "batch_isend_irecv",
+           "reduce",
+           "ReduceOp",
+           "new_group",
+           "get_group",
+           "is_initialized",
+           "destroy_process_group",
+           "scatter",
+           "split",
+           "wait",
+           "all_reduce",
+           "barrier",
+           "broadcast",
+           "recv",
+           "irecv",
+           "reduce_scatter",
+           "send",
+           "isend",
+           "P2POp",
+           ]
