+++ python/setup.py.in
@@ -316,6 +316,7 @@ packages=['paddle',
           'paddle.distributed.passes',
           'paddle.distributed.models',
           'paddle.distributed.models.moe',
+          'paddle.distributed.communication',
           'paddle.framework',
           'paddle.jit',
           'paddle.jit.dy2static',
